reorganize Test file
